third_party/WebKit/LayoutTests/platform/win/external/wpt/html/webappapis/scripting/events/event-handler-onauxclick-expected.txt

This is a testharness.js-based test.
PASS onauxclick is on the appropriate locations for GlobalEventHandlers 
PASS The default value of onauxclick is always null 
PASS The onauxclick content attribute must be compiled into the onauxclick property 
PASS The onauxclick content attribute must execute when an event is dispatched 
PASS dispatching an auxclick event must trigger element.onauxclick 
Harness: the test ran to completion.
